M<enu Ui, timer, score

+++ Assets/Scripts/Antek/SO/SOFloat.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 [CreateAssetMenu]
 public class SOFloat : ScriptableObject
 {
-   public float value;
+   public int value;
 }
